Separate out impersonation action.
